+++ src/clld_corpus_plugin/interfaces.py
@@ -23,3 +23,8 @@ class IWordform(Interface):
     """
     form interface
     """
+
+class ISpeaker(Interface):
+    """
+    speaker interface
+    """